Refactored user api Pulled out some of the longer sections of the function to a new module and added more error handling relates to #58, #71
